pyproject.toml
[tool.poetry]
name = "ai-career-copilot"
version = "0.1.0"
description = "AI-powered career assistant for resume analysis and job matching"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.28.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
sentence-transformers = "^2.2.0"
openai = "^1.0.0"
spacy = "^3.7.0"
transformers = "^4.30.0"
torch = "^2.0.0"
PyMuPDF = "^1.23.0"
python-docx = "^0.8.11"
pdfminer-six = "^20221105"
docx2txt = "^0.8"
beautifulsoup4 = "^4.12.0"
selenium = "^4.15.0"
requests = "^2.31.0"
lxml = "^4.9.0"
pymongo = "^4.5.0"
scikit-learn = "^1.3.0"
nltk = "^3.8.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
flake8 = "^6.0.0"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]